Correctly escape displayed json values and put response in window.API_RESPONSE
Summary: add window.API_RESPONSE variable which contains the preformated api response and some helpful tips on the homepage about it. The original rationale behind escaping the displayed json was that it's probably not so unfathomable that a person using the explorer would try to parse the #response_body variable, but it's actually not encoded correctly. In any way, to discourage that (and to encourage actually poking of the api) i'm exposing the response json in a variable which gets set after each call is made.

Test Plan:
$.parseJSON($("#response_body").text()) works.
additionally, and more usefully, window.API_RESPONSE is available for perusing rather than mucking with #response_body.

Reviewers: kamens

Reviewed By: kamens

Differential Revision: http://phabricator.khanacademy.org/D975