Template and CSS styles for API Explorer
Summary: templates and css styles for api explorer. Most of the client side code is generated by javascript. Stylus is a superset of CSS so any valid css can be used in the stylesheet. I can replace it with raw css if that's desired, but the syntax is so much nicer.

Reviewers: alpert, joel

Reviewed By: alpert

CC: ben

Maniphest Tasks: T1210

Differential Revision: http://phabricator.khanacademy.org/D3007