Fix for google app engine truncating too long headers
Summary:
This is an ugly hack which is necessary due to google app engine design. It is not stated anywhere directly but google app engine trims headers which are more than 498 bytes long (this number might be wrong). I have managed to dig out an old issue that mentions this but it is nowhere in official documentation. Anyway this is consistent with the behaviour I have observed where X-Original-Headers gets removed from the response. If you have any other suggestions how this might be avoided I will be more than happy to implement them. Also if you want to file a bug report with google then you're more than welcome.

Remove function included for testing

Test Plan:
headers are present in the api response.

<see below>

Reviewers: alpert, csilvers

Reviewed By: alpert

CC: ben

Maniphest Tasks: T1210

Differential Revision: http://phabricator.khanacademy.org/D3098