Make the test client show the response body for HTTP errors
Summary:
Previously, only the title of an error response (e.g. "Unauthorized" for 401)
was displayed when there was an HTTP error. I'm introducing a new error and I
want the message to be more visible, so this at least makes it visible in the
example client.

Test Plan:
Register a consumer in dev_appserver.
Run /api/v1/user/students. It fails with 401 because the consumer isn't
anointed, and the response message is displayed in the client (previously, it
just said "Unauthorized").

Reviewers: alpert

Reviewed By: alpert

Subscribers: ben

Differential Revision: http://phabricator.khanacademy.org/D12243