Add a sample PHP client
This uses the oauth-php library to go through a browser-based authentication
flow. I also included the oauth-php library directly as a dependency.

Test Plan:
Visit localhost:8001/ka_client.php, click the login button, log in, it redirects
to a page with a form for making a GET request. Run the default request and
gives back my user data.

Auditors: marcos